build: add golangci-lint to the project
Initial golangci-lint config is added to the project. The Makefile is
configured to download golangci-lint binary if missing. The Makefile now
contains `lint` and `lint-warn` commands both running the linter, `lint-warn`
does not throw an error.
`fmt` is also updated in Makefile to call `go mod tidy` and golangci-lint with
fix parameter.
docker build-image is also updated to use the same version as in Makefile via
docker build args.

Signed-off-by: Umar Hussain <umarhussain.work@gmail.com>